+++ src/mechir/data/loader/cat.py
@@ -1,4 +1,4 @@
-from . import BaseCollator
+from .base import BaseCollator
 
 
 class CatDataCollator(BaseCollator):
